+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* The resfinder.tsv and pointfinder.tsv outputs now contain a Notes column.
 * Updated the help description of the --mlst-scheme parameter to include a more useful link for available schemas.
 * Switched to only officially supporting Python 3.7+ due to recent incompatibilities with Python 3.6 and some Python packages (numpy, biopython, and others).
 * Adds the ability to handle "complex" pbp5 mutations. When appriopriate many pbp5 point mutations will be reported as a single mutation.